add authorization to thread endpoints

+++ server/graphql/resolvers/thread.js
@@ -32,27 +32,32 @@ export default {
     }
   },
   Mutation: {
-    createThread: async (parent, args, { models }) => {
+    createThread: async (parent, args, { models, user }) => {
       try {
-        return await models.Thread.create()
+        console.log(user.id)
+        if (!user) throw new Error(`Anauthorized`)
+        return await models.Thread.create({ userId: user.id })
       } catch (e) {
         console.log(e)
         throw e
       }
     },
-    updateThread: async (parent, { id, detail }, { models }) => {
+    updateThread: async (parent, { id, detail }, { models, user }) => {
       try {
+        if (!user) throw new Error(`Anauthorized`)
         const thread = await models.Thread.findOne({ where: { id } })
         if (!thread) throw new Error(`Not found thread with ${id}`)
+        if (user.id !== thread.userId) throw new Error(`Anauthorized`)
         return await thread.update({ detail })
       } catch (e) {
         console.log(e)
         throw e
       }
     },
-    deleteThread: async (parent, { id }, { models }) => {
+    deleteThread: async (parent, { id }, { models, user }) => {
       try {
-        await models.Thread.destroy( { where: { id } })
+        if (!user) throw new Error(`Anauthorized`)
+        await models.Thread.destroy( { where: { id, userId: user.id } })
         return true
       } catch (e) {
         console.log(e)
